Solution for two sum problem.
The problem input is an unsorted array and a target number. The goal is to
determine the indices in the array that adds up to the target number.
The key to the problem is to solve it in O(N) time complexity.
The insight is to store the expected difference between the target and every
number encountered in a hash table along with the slot number. If the
difference is encountered along the scan, then we can return the stored slot
number and the current slot number as the result!