Rotate a caller-specified vector/array by a constant number of steps.
For e.g.,
./a.out 2 1 2 3 4 5 6
5 6 1 2 3 4
where 2 is the number of steps and {1, 2, 3, 4, 5, 6} is the vector to be
rotated.
The idea here is to rotate a vector by a certain number of times.
The first step here is to split the incoming vector into 2 subvectors
based on the step number,
in this case {1, 2, 3, 4} and {5, 6}.
Then we reverse each of these subvectors to form {4, 3, 2, 1} and {6, 5}
and we append them together to yield {4, 3, 2, 1, 6, 5}
and then reverse the appended vector to yield the desired answer {5, 6, 1, 2, 3, 4}