Program to compute the length of the maximum substring with no repeated characters.
THe key here is to scan all characters no more than once. We use a set to keep
track of duplicate/previously encountered characters. We use a string auxilary
variable to slide through the characters. When a duplicate is encountered, we
do a sliding window into the string encountered so far and remove the first
character from the set as well as from the substring.