Program to print atmost 2 unique characters max. substring.
The idea here is to scan through each character and track if we have seen more
than 2 unique characters. If we haven't exceeded we keep appending to a string
and move on. If this character caused us to exceed the limit, then we remember
the max. that we have seen thus far and reset the string to point to the
character prior to the current one.