Added program to serialize and deserialize a binary tree (or bst).
This program serializes a binary tree structure into a string by effectively
doing a breadth-first-search walk along the tree. There is an equivalent
deserializer function that takes the serialized string and constructs the
original tree back.

Added various test cases to construct and validate the output as well as
assorted utility routines to do a bfs search traversal.

./build/binarytreeencode
4
8
10
12
14
20
22
45
[20 8 22 4 12 null 45 null null 10 14 null null null null null null ]
4
8
10
12
14
20
22
45
[0] 5
[1] 3
[1] 2
[2] 4
[2] 10
[3] 1
[3] 8
[5 3 2 4 null null 10 1 null null 8 null null null null ]
[0] 5
[1] 3
[1] 2
[2] 4
[2] 10
[3] 1
[3] 8