Overlapping range problem detection.
The algorithm for detection of minimum number of ranges in a target range is as
follows.
- Create a new set sorted in descending order of overlap with the target range.
- Greedily add the overlapping ranges except,
  - the range already overlaps with what we have so far.
  - the range doesn't intersect with what we have so far.
  - the new range breaks the left or right clamp.
Once added, extend the range that we have so far and set the right or left
clamp accordingly.
If both left and right clamps have been set, we are done and we can return the
set of ranges added. If not and we terminate the iteration then we throw an
exception.

./build/overlapRange
1 14:1 4, 3 9, 9 12, 11 14,
  3 13:3 9, 9 12, 11 14,
  12 17:11 14, 14 20,
  -4 -1:Not possible
  4 7:3 9,
  4 17:3 9, 9 12, 11 14, 14 20,
  New Problem
  3 13:1 10, 7 14,
  -3 13:-5 1, 1 10, 7 14,