Range sum problem.
The problem is to implement a range sum for a single dimension array.
The first approach is to use a cumulative sum one for each element in the
array. However, this is linear in time complexity.
The second approach is to implement what is termed as a segment tree,
where we recursively build the tree from the array by sub-dividing into 2 equal
halves.